Port ubiquity kde frontend to be just Qt
Ubiquity KDE frontend was written as a KApplication. This, combined
with the fact that ubiquity is running as root using sudo and the
fact that KApplication does behind-the scenes calls to dbus *session*
bus during initialization, where always causing problems. So far
ubiquity managed to "cheat" its way out of this by hiding its
"rootliness" through hacks.

Because of CVE-2012-3524 though, libdbus now has a solid way to
know if its caller is really seteuid so these tricks are no more.

This brought us LP: #1055967, i.e., ubiquity kde broke by just
a *security update* during the *stable series*. So nice.

Luckily this motivated our cool hacker Jonathan Riddell to do
a rewrite of this frontend as a QApplication (this is what this
patch is about).

Since Qt doesn't do session bus calls behind our back, this should
solve this type of problem once and for all.