Fix localization bug of timezone selection screen
By default, ubiquity was starting with "en" locale. Thus when you
were going to the timezone selection screen, you would see the
time like

10.03.28 PM

in the tooltip over the map, even with the Greek interface.

In contrast, in squibby we start ubiquity using the Greek locale.

This exposed a bug that this tooltip has, that it uses latin-1
instead of UTF-8, so it was showing the time and then gibberish
instead of πμ or μμ.

Thıs patch fixes the bug.